+++ packages/hex/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-- **Last updated**: 2025-11-13T12:38:59Z
+- **Last updated**: 2025-12-03T22:43:13Z
 - **Generator**: [thi.ng/monopub](https://thi.ng/monopub)
 
 All notable changes to this project will be documented in this file.