+++ packages/hiccup-canvas/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-- **Last updated**: 2025-10-25T22:30:00Z
+- **Last updated**: 2025-12-03T22:43:13Z
 - **Generator**: [thi.ng/monopub](https://thi.ng/monopub)
 
 All notable changes to this project will be documented in this file.
@@ -10,6 +10,27 @@ See [Conventional Commits](https://conventionalcommits.org/) for commit guidelin
 
 **Note:** Unlisted _patch_ versions only involve non-code or otherwise excluded changes
 and/or version bumps of transitive dependencies.
+
+### [3.2.3](https://github.com/thi-ng/umbrella/tree/@thi.ng/hiccup-canvas@3.2.3) (2025-12-03)
+
+#### ♻️ Refactoring
+
+- update for-loops (use const where possible) ([5ceaf1a](https://github.com/thi-ng/umbrella/commit/5ceaf1a))
+
+## [3.2.0](https://github.com/thi-ng/umbrella/tree/@thi.ng/hiccup-canvas@3.2.0) (2025-11-19)
+
+#### 🚀 Features
+
+- [#554](https://github.com/thi-ng/umbrella/issues/554), add `__dpr` root attrib support ([74e4618](https://github.com/thi-ng/umbrella/commit/74e4618))
+  - add support for `__dpr` control attrib in root shape/group
+    to set/adjust canvas pixel size (and scale context) to given
+    devicePixelRatio (see [@thi.ng/canvas](https://github.com/thi-ng/umbrella/tree/main/packages/canvas) adaptDPI() for details)
+  - minor update handling of other control attribs
+  - add [@thi.ng/canvas](https://github.com/thi-ng/umbrella/tree/main/packages/canvas) dependency
+
+#### ♻️ Refactoring
+
+- minor internal updates ([8312695](https://github.com/thi-ng/umbrella/commit/8312695))
 
 ### [3.1.31](https://github.com/thi-ng/umbrella/tree/@thi.ng/hiccup-canvas@3.1.31) (2025-01-14)