+++ packages/imago/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-- **Last updated**: 2025-10-25T22:30:00Z
+- **Last updated**: 2025-12-03T22:43:13Z
 - **Generator**: [thi.ng/monopub](https://thi.ng/monopub)
 
 All notable changes to this project will be documented in this file.
@@ -10,6 +10,18 @@ See [Conventional Commits](https://conventionalcommits.org/) for commit guidelin
 
 **Note:** Unlisted _patch_ versions only involve non-code or otherwise excluded changes
 and/or version bumps of transitive dependencies.
+
+### [1.5.4](https://github.com/thi-ng/umbrella/tree/@thi.ng/imago@1.5.4) (2025-12-03)
+
+#### ♻️ Refactoring
+
+- update for-loops (use const where possible) ([5ceaf1a](https://github.com/thi-ng/umbrella/commit/5ceaf1a))
+
+## [1.5.0](https://github.com/thi-ng/umbrella/tree/@thi.ng/imago@1.5.0) (2025-11-17)
+
+#### 🚀 Features
+
+- add `maxsize()` operator ([d8484ff](https://github.com/thi-ng/umbrella/commit/d8484ff))
 
 ## [1.4.0](https://github.com/thi-ng/umbrella/tree/@thi.ng/imago@1.4.0) (2025-07-25)