+++ packages/mime/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-- **Last updated**: 2025-10-25T22:30:00Z
+- **Last updated**: 2025-12-03T22:43:13Z
 - **Generator**: [thi.ng/monopub](https://thi.ng/monopub)
 
 All notable changes to this project will be documented in this file.
@@ -10,6 +10,22 @@ See [Conventional Commits](https://conventionalcommits.org/) for commit guidelin
 
 **Note:** Unlisted _patch_ versions only involve non-code or otherwise excluded changes
 and/or version bumps of transitive dependencies.
+
+### [2.8.2](https://github.com/thi-ng/umbrella/tree/@thi.ng/mime@2.8.2) (2025-12-03)
+
+#### ♻️ Refactoring
+
+- update for-loops (use const where possible) ([5ceaf1a](https://github.com/thi-ng/umbrella/commit/5ceaf1a))
+
+## [2.8.0](https://github.com/thi-ng/umbrella/tree/@thi.ng/mime@2.8.0) (2025-11-25)
+
+#### 🚀 Features
+
+- update conversion tool & MIME type overrides ([266a191](https://github.com/thi-ng/umbrella/commit/266a191))
+- update MIME index, add overrides ([44a0a2e](https://github.com/thi-ng/umbrella/commit/44a0a2e))
+  - add typescript MIME types
+  - update preferred .m4a, .ts, .wav mappings
+  - update preferrd `video/quicktime` mapping (.mov)
 
 ## [2.7.0](https://github.com/thi-ng/umbrella/tree/@thi.ng/mime@2.7.0) (2025-01-29)